데이터 접근 방식 수정 / Bread.id 모델 수정
1. 데이터의 접근을 한 곳에서 하기 위해, BreadDAO를 싱글톤 객체로 만들고 NSManagedObject를 들고 있게 수정.
2. 뷰와 데이터를 분리하기 위해, BreadListVC는 Controller를 통해서만 스냅샷용 아이템을 받도록 수정.
3. BreadListController는 BreadDAO의 싱글톤 객체에서 Bread 인스턴스를 받아서 아이템을 만듦
4. BreadVC는 처음에 컨트롤러 없이 구현했기에 Bread 객체를 외부로부터 주입받는 방식 유지. 이때문에 BreadListVC에서 BreadDAO에 바로 접근할 때가 있음.
5. Bread별 명확한 구분을 위해 Bread.id를 UUID에서 autoIncrease 방식으로 변경.

+++ MemoryBread/Models/Bread/BreadDAO.swift
@@ -8,11 +8,21 @@
 import Foundation
 
 final class BreadDAO {
+    static let `default` = BreadDAO()
+    
+    private lazy var breads: [Bread] = {
+        return fetchAll()
+    }()
+    
+    var allBreads: [Bread] {
+        return breads
+    }
+    
     private func Log(title: String, error: Error) {
         NSLog("\(title) failed. Error=\(error)")
     }
     
-    func fetchAll() -> [Bread] {
+    private func fetchAll() -> [Bread] {
         let request = Bread.fetchRequest()
         let sortDescriptor = NSSortDescriptor(key: "touch", ascending: false)
         request.sortDescriptors = [sortDescriptor]
@@ -39,17 +49,28 @@ final class BreadDAO {
         return true
     }
     
-    func delete(_ bread: Bread) {
+    @discardableResult
+    func delete(_ bread: Bread) -> Bool {
+        guard let index = breads.firstIndex(of: bread) else { return false }
+        breads.remove(at: index)
         AppDelegate.viewContext.delete(bread)
+        return true
     }
     
     func create() -> Bread {
-        return Bread(touch: Date.now,
+        let bread = Bread(touch: Date.now,
                           directoryName: "임시 디렉토리",
                           title: "새로운 암기빵",
                           content: "",
                           separatedContent: [],
                           filterIndexes: Array(repeating: [], count: FilterColor.count))
+        breads.insert(bread, at: 0)
+        return bread
+    }
+    
+    func bread(at index: Int) -> Bread? {
+        guard breads.count > 0 && index < breads.count else { return nil }
+        return breads[index]
     }
 }
